pyproject.toml
[tool.poetry]
name = "repl_nix_Doge-Memer"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
discord = "^2.0.0"
Flask = "^2.2.2"
pymongo = "^4.2.0"
python-dotenv = "^0.21.0"
rapidfuzz = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
